Updated answer in question 133 as Azure is decommissioning that option
Hi Daniel! I am loving the AZ-204 repository for my exam prep! I was doing a little extra digging for question 133 and I found that one of the selected answers "*Multi-step web*" is being decommissioned: https://azure.microsoft.com/en-ca/updates/retirement-notice-transition-to-custom-availability-tests-in-application-insights/ so I don't think it should remain as a right option or even one of the available options for this question.

Let me know your thoughts! 

Thanks,